pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nucsys-agent"
version = "0.2.0"
description = "Topology + sizing agent that outputs Alchemy-style JSON graphs for nuclear systems."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.5",
  "pydantic-settings>=2.2",
  "httpx>=0.25",
  "tenacity>=8.0",
  "pyyaml>=6.0",
  "networkx>=3.0",
  "numpy>=1.24",
  "scipy>=1.10",
  "orjson>=3.9",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "iapws>=1.5; platform_system!='Windows'",  # water/steam props (pure python)
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.3",
]

[project.scripts]
nucsys-agent = "nucsys_agent.cli:main"

[tool.setuptools]
packages = {find = {where = ["."], include = ["nucsys_agent*"]}}
include-package-data = true

[tool.setuptools.package-data]
nucsys_agent = ["data/*.yaml", "data/cards/*.yaml", "schemas/*.json"]

[tool.ruff]
line-length = 100
